demo-ios: fix Combine import + Int32 cast for build under Xcode 26
Two compile errors surfaced when adding the source to a fresh iOS App
project under Xcode 26.4.1:

  1. @Published / ObservableObject required an explicit `import Combine`.
     SwiftUI no longer re-exports Combine implicitly in iOS 17+ targets.

  2. `config.iterations = iterations` failed because TrainingConfig.iterations
     is Int32 (bridges to C `int`) but SwiftUI Stepper produces Int. Added
     `Int32(iterations)` cast.

+++ demo-ios/Sources/IpadTestApp/IpadTestApp.swift
@@ -1,5 +1,6 @@
 import SwiftUI
 import UIKit
+import Combine
 import QuartzCore
 import Msplat
 
@@ -71,7 +72,7 @@ final class Engine: ObservableObject {
             autoreleasepool {
                 let dataset = GaussianDataset(path: datasetPath)
                 var config = TrainingConfig()
-                config.iterations = iterations
+                config.iterations = Int32(iterations)
                 config.numDownscales = 0
                 config.bgColor = (0, 0, 0)
                 let trainer = GaussianTrainer(dataset: dataset, config: config)
